rework pipeline.yaml
